cast .tv_sec to avoid warning on DARWIN

+++ contrib/uuid/gen_uuid.c
@@ -405,7 +405,8 @@ try_again:
 		rewind(state_f);
 		len = fprintf(state_f, 
 			      "clock: %04x tv: %016lu %08lu adj: %08d\n",
-			      clock_seq, last.tv_sec, last.tv_usec, adjustment);
+			      /* cast to unsigned long to avoid warning (int on DARWIN) */
+			      clock_seq, last.tv_sec, (unsigned long) last.tv_usec, adjustment); 
 		fflush(state_f);
 		if (ftruncate(state_fd, len) < 0) {
 			fprintf(state_f, "                   \n");
